feat(but-workspace): add function to move multiple commits
